Fix a memory error in AudioBuffer
std::align takes the pointer to the buffer as a reference and
modifies it; this means that _dataBase here was being modified
on unaligned mallocs which would make the subsequent free also
fail. Without this change, valgrind would spit pages of errors
and clap-info would crash on exit. With it, valgrind is clean
running the plugins in clap-info